Begin implementation of type configuration
AWS recently introduced type configurations for registered cloudformation providers. This changes implementation pretty significantly. This is the first pass at making this provider work with type configuration.

Its currently unclear how testing works with type configuration and some details are still returning as blank that are needed for create and read operations so this provider is still a WIP